Refactor the code in PathUtil
