+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/BallerinaTextDocumentService.java
@@ -23,7 +23,7 @@ import org.ballerinalang.formatter.core.Formatter;
 import org.ballerinalang.formatter.core.FormatterException;
 import org.ballerinalang.langserver.codelenses.CodeLensUtil;
 import org.ballerinalang.langserver.codelenses.LSCodeLensesProviderHolder;
-import org.ballerinalang.langserver.common.utils.CommonUtil;
+import org.ballerinalang.langserver.common.utils.PathUtil;
 import org.ballerinalang.langserver.commons.BallerinaDefinitionContext;
 import org.ballerinalang.langserver.commons.CodeActionContext;
 import org.ballerinalang.langserver.commons.CompletionContext;
@@ -163,7 +163,7 @@ class BallerinaTextDocumentService implements TextDocumentService {
         return CompletableFutures.computeAsync((cancelChecker) -> {
             try {
                 HoverContext context = ContextBuilder.buildHoverContext(
-                        CommonUtil.convertUriSchemeFromBala(params.getTextDocument().getUri()),
+                        PathUtil.convertUriSchemeFromBala(params.getTextDocument().getUri()),
                         this.workspaceManagerProxy.get(),
                         this.serverContext, params.getPosition(),
                         cancelChecker);
@@ -185,7 +185,7 @@ class BallerinaTextDocumentService implements TextDocumentService {
     public CompletableFuture<SignatureHelp> signatureHelp(SignatureHelpParams params) {
         return CompletableFutures.computeAsync((cancelChecker) -> {
             String uri = params.getTextDocument().getUri();
-            Optional<Path> sigFilePath = CommonUtil.getPathFromURI(uri);
+            Optional<Path> sigFilePath = PathUtil.getPathFromURI(uri);
 
             // Note: If the path does not exist, then return early and ignore
             if (sigFilePath.isEmpty()) {
@@ -221,7 +221,7 @@ class BallerinaTextDocumentService implements TextDocumentService {
         return CompletableFutures.computeAsync((cancelChecker) -> {
             try {
                 BallerinaDefinitionContext defContext = ContextBuilder.buildDefinitionContext(
-                        CommonUtil.convertUriSchemeFromBala(params.getTextDocument().getUri()),
+                        PathUtil.convertUriSchemeFromBala(params.getTextDocument().getUri()),
                         this.workspaceManagerProxy.get(),
                         this.serverContext,
                         params.getPosition(),
@@ -247,7 +247,7 @@ class BallerinaTextDocumentService implements TextDocumentService {
             try {
                 String fileUri = params.getTextDocument().getUri();
                 ReferencesContext context = ContextBuilder.buildReferencesContext(
-                        CommonUtil.convertUriSchemeFromBala(fileUri),
+                        PathUtil.convertUriSchemeFromBala(fileUri),
                         this.workspaceManagerProxy.get(fileUri),
                         this.serverContext,
                         params.getPosition(),
@@ -263,9 +263,9 @@ class BallerinaTextDocumentService implements TextDocumentService {
                             Path filePath = ReferencesUtil.getPathFromLocation(module, location);
                             String uri = filePath.toUri().toString();
                             // If path is readonly, change the URI scheme
-                            if (CommonUtil.isWriteProtectedPath(filePath)) {
+                            if (PathUtil.isWriteProtectedPath(filePath)) {
                                 try {
-                                    uri = CommonUtil.getBalaUriForPath(serverContext, filePath);
+                                    uri = PathUtil.getBalaUriForPath(serverContext, filePath);
                                 } catch (URISyntaxException e) {
                                     this.clientLogger.logError(LSContextOperation.TXT_REFERENCES,
                                             "Failed to convert path to bala URI", e,
@@ -295,7 +295,7 @@ class BallerinaTextDocumentService implements TextDocumentService {
     documentSymbol(DocumentSymbolParams params) {
         return CompletableFutures.computeAsync((cancelChecker) -> {
             String fileUri = params.getTextDocument().getUri();
-            Optional<Path> docSymbolFilePath = CommonUtil.getPathFromURI(fileUri);
+            Optional<Path> docSymbolFilePath = PathUtil.getPathFromURI(fileUri);
 
             // Note: If the path does not exist, then return early and ignore
             if (docSymbolFilePath.isEmpty()) {
@@ -357,7 +357,7 @@ class BallerinaTextDocumentService implements TextDocumentService {
             }
 
             String fileUri = params.getTextDocument().getUri();
-            Optional<Path> docSymbolFilePath = CommonUtil.getPathFromURI(fileUri);
+            Optional<Path> docSymbolFilePath = PathUtil.getPathFromURI(fileUri);
 
             // Note: If the path does not exist, then return early and ignore
             if (docSymbolFilePath.isEmpty()) {
@@ -522,7 +522,7 @@ class BallerinaTextDocumentService implements TextDocumentService {
         String fileUri = params.getTextDocument().getUri();
         try {
             DocumentServiceContext context = ContextBuilder.buildDocumentServiceContext(
-                    CommonUtil.convertUriSchemeFromBala(fileUri),
+                    PathUtil.convertUriSchemeFromBala(fileUri),
                     this.workspaceManagerProxy.get(fileUri),
                     LSContextOperation.TXT_DID_OPEN, this.serverContext);
             this.workspaceManagerProxy.didOpen(params);
@@ -544,7 +544,7 @@ class BallerinaTextDocumentService implements TextDocumentService {
         try {
             // Update content
             DocumentServiceContext context = ContextBuilder.buildDocumentServiceContext(
-                    CommonUtil.convertUriSchemeFromBala(fileUri),
+                    PathUtil.convertUriSchemeFromBala(fileUri),
                     this.workspaceManagerProxy.get(fileUri),
                     LSContextOperation.TXT_DID_CHANGE,
                     this.serverContext);
@@ -566,7 +566,7 @@ class BallerinaTextDocumentService implements TextDocumentService {
         String fileUri = params.getTextDocument().getUri();
         try {
             DocumentServiceContext context = ContextBuilder.buildDocumentServiceContext(
-                    CommonUtil.convertUriSchemeFromBala(fileUri),
+                    PathUtil.convertUriSchemeFromBala(fileUri),
                     this.workspaceManagerProxy.get(fileUri),
                     LSContextOperation.TXT_DID_CLOSE,
                     this.serverContext);
