+++ language-server/modules/langserver-core/src/test/java/org/ballerinalang/langserver/definition/DefinitionTest.java
@@ -23,6 +23,7 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import org.ballerinalang.langserver.common.utils.CommonUtil;
+import org.ballerinalang.langserver.common.utils.PathUtil;
 import org.ballerinalang.langserver.commons.capability.InitializationOptions;
 import org.ballerinalang.langserver.util.FileUtils;
 import org.ballerinalang.langserver.util.TestUtil;
@@ -100,7 +101,7 @@ public class DefinitionTest {
 
         String fileUri = sourcePath.toUri().toString();
         if (withBalaScheme) {
-            fileUri = CommonUtil.getUriForPath(sourcePath, CommonUtil.URI_SCHEME_BALA);
+            fileUri = PathUtil.getUriForPath(sourcePath, CommonUtil.URI_SCHEME_BALA);
         }
 
         byte[] encodedContent = Files.readAllBytes(sourcePath);
@@ -216,7 +217,7 @@ public class DefinitionTest {
             URI  uri = new URI(fileUri);
             Assert.assertEquals(uri.getScheme(), getExpectedUriScheme(),
                     String.format("Expected %s: URI scheme", getExpectedUriScheme()));
-            fileUri = CommonUtil.convertUriSchemeFromBala(fileUri);
+            fileUri = PathUtil.convertUriSchemeFromBala(fileUri);
             uri = new URI(fileUri);
             Assert.assertEquals(uri.getScheme(), CommonUtil.URI_SCHEME_FILE,
                     "Expected file URI scheme after conversion");
